feat: merge redirect page into login page
