+++ apps/renderer/src/lib/auth.ts
@@ -1,7 +1,7 @@
 import { IN_ELECTRON, WEB_URL } from "@follow/shared/constants"
 import { signIn } from "@hono/auth-js/react"
 
-export const LOGIN_CALLBACK_URL = `${WEB_URL}/redirect?app=follow`
+export const LOGIN_CALLBACK_URL = `${WEB_URL}/login`
 export type LoginRuntime = "browser" | "app"
 export const loginHandler = (provider: string, runtime: LoginRuntime = "app") => {
   if (IN_ELECTRON) {
