+++ apps/server/client/components/ui/user-avatar.tsx
@@ -1,9 +1,10 @@
 import { useWhoami } from "@client/atoms/user"
 import { Avatar, AvatarFallback, AvatarImage } from "@follow/components/ui/avatar/index.jsx"
 import { getBackgroundGradient } from "@follow/utils/color"
+import { cn } from "@follow/utils/utils"
 import { useMemo } from "react"
 
-export const UserAvatar = () => {
+export const UserAvatar = ({ className }: { className?: string }) => {
   let user = useWhoami()
 
   const colors = useMemo(
@@ -31,7 +32,12 @@ export const UserAvatar = () => {
   const { name, image } = user
 
   return (
-    <div className="flex h-20 items-center justify-center gap-8 px-10 py-4 text-2xl font-medium text-zinc-600 dark:text-zinc-300">
+    <div
+      className={cn(
+        "flex h-20 items-center justify-center gap-8 px-10 py-4 text-2xl font-medium text-zinc-600 dark:text-zinc-300",
+        className,
+      )}
+    >
       <Avatar className="border">
         <AvatarImage
           className="aspect-square size-full duration-200 animate-in fade-in-0"
