+++ vercel.json
@@ -21,10 +21,6 @@
       "destination": "https://follow-external-ssr.vercel.app/login"
     },
     {
-      "source": "/redirect",
-      "destination": "https://follow-external-ssr.vercel.app/redirect"
-    },
-    {
       "source": "/external-dist/:path*",
       "destination": "https://follow-external-ssr.vercel.app/external-dist/:path*"
     },
